References Oppevara/oppevara#61, changed semantics file.
Changed all texts to estonian (will not provide translations at this
point as all the text within the library itself are only in estonian).
Changed default value and minimum value for number of attempts. Also
added an explanations a out what it affects (although the fature has not
been implemented yet).